ARM: Kirkwood: Allow use of pinctrl
We now have a pinctrl driver for arch-mvebu. It supports kirkwood out
of the box. Allow ARCH_KIRKWOOD to select it, and make ARCH_KIRKWOOD
select it. This then allows the convertion of board-*.c files away
from MPP to pinctrl in DT, before they get moved over into mach-mvebu.

Signed-off-by: Andrew Lunn <andrew@lunn.ch>